Implemented complete Employee CRUD operations using Spring Boot and Spring Data JPA

+++ src/main/java/com/aaruu/ems/service/EmployeeServiceImpl.java
@@ -51,14 +51,34 @@ public class EmployeeServiceImpl implements EmployeeService {
 	@Override
 	public Employee getEmployeeById(Integer id) {
 		return employeeRepository.findById(id).orElse(null);
+		// here we can also write orElseThrow(()->new RuntimeException("employee not
+		// found"));
 	}
 	// Find employee by id. If found, return employee. Otherwise return null.
 	// findById() returns an Optional<Entity>. It helps avoid NullPointerException
 	// and provides a safe way to handle missing records.
 
+	// In REST APIs, returning 404 Not Found is preferred because the requested
+	// resource does not exist.
+//    Returning 200 OK with null can be misleading
+//First I would check whether the record exists using findById(). If the record is found, I would update it. Otherwise, 
+//	I would return a 404 Not Found response
 	@Override
-	public Employee updateEmployee(Employee employee) {
-		return employeeRepository.save(employee);
+	public Employee updateEmployee(Integer id, Employee employee) {
+
+		Employee existingEmployee = employeeRepository.findById(id).orElse(null);
+		if (existingEmployee == null) {
+			return null;
+//throw new RuntimeException("Employee not found");
+		}
+
+		existingEmployee.setFirstName(employee.getFirstName());
+		existingEmployee.setLastName(employee.getLastName());
+		existingEmployee.setEmail(employee.getEmail());
+		existingEmployee.setDepartment(employee.getDepartment());
+		existingEmployee.setSalary(employee.getSalary());
+
+		return employeeRepository.save(existingEmployee);
 	}
 	// JpaRepository uses the save() method for both insert and update operations.
 	// If the entity already exists with a valid primary key, save() performs an
@@ -66,6 +86,10 @@ public class EmployeeServiceImpl implements EmployeeService {
 
 	@Override
 	public void deleteEmployee(Integer id) {
+		Employee employee = employeeRepository.findById(id).orElse(null);
+		if (employee == null) {
+			return;
+		}
 		employeeRepository.deleteById(id);
 	}
 	// JpaRepository provides the deleteById() method, which deletes a record based
